requirements.txt
torch>=1.7.0
transformers>=4.2.2
python-telegram-bot>=13.1